requirements.txt
Flask==0.9
Flask-OAuth==0.12
Flask-SQLAlchemy==0.16
Flask-Script==0.5.3
Jinja2==2.6
SQLAlchemy==0.8.0
Werkzeug==0.8.3
argparse==1.2.1
gevent==0.13.8
gevent-socketio==0.3.6
gevent-websocket==0.3.6
greenlet==0.4.0
httplib2==0.8
oauth2==1.5.211
psycogreen==1.0
redis==2.7.2
wsgiref==0.1.2
